+++ src/borg/legacy/remote.py
@@ -25,6 +25,8 @@ from ..helpers import prepare_subprocess_env, ignore_sigint
 from ..fslocking import LockTimeout, NotLocked, NotMyLock, LockFailed
 from ..logger import create_logger, borg_serve_log_queue
 from ..helpers import msgpack
+from ..helpers import safe_encode
+from ..manifest import Manifest, NoManifestError
 from .repository import LegacyRepository
 from ..repository import Repository, StoreObjectNotFound
 from ..version import parse_version, format_version
@@ -37,6 +39,21 @@ BORG_VERSION = parse_version(__version__)
 MSGID, MSG, ARGS, RESULT, LOG = "i", "m", "a", "r", "l"
 
 MAX_INFLIGHT = 100
+
+# the RPC methods whose results are bytes (or tuples of bytes)
+BYTES_RESULT_METHODS = ("open", "get", "list", "scan", "load_key")
+
+
+def bytes_from_borg1(value):
+    """Convert a result of a borg 1.x server back to bytes (or a tuple of bytes).
+
+    borg 1.x packs bytes into the msgpack "raw" type, which we unpack to str (see helpers.msgpack).
+    """
+    if isinstance(value, str):
+        return safe_encode(value)
+    if isinstance(value, (list, tuple)):
+        return tuple(bytes_from_borg1(v) for v in value)
+    return value
 
 
 class ConnectionClosed(Error):
@@ -224,6 +241,8 @@ class LegacyRemoteRepository:
         self.shutdown_time = None
         self.unpacker = get_limited_unpacker("client")
         self.server_version = None  # we update this after server sends its version
+        # borg 1.x servers do not have the info, close and get_manifest methods and they send bytes as str.
+        self.borg1_server = False
         self.p = None
         self._args = args
         if self.location.proto == "ssh":
@@ -273,6 +292,7 @@ class LegacyRemoteRepository:
                 self.server_version = version["server_version"]
             else:
                 raise Exception("Server insisted on using unsupported protocol version %s" % version)
+            self.borg1_server = self.server_version < parse_version("2.0.0a1")
 
             self.id = self.open(
                 path=self.location.path,
@@ -282,8 +302,11 @@ class LegacyRemoteRepository:
                 exclusive=exclusive,
                 v1_legacy=True,  # make remote use LegacyRepository
             )
-            info = self.info()
-            self.version = info["version"]
+            if self.borg1_server:
+                self.version = 1  # borg 1.x servers only serve v1 repositories.
+            else:
+                info = self.info()
+                self.version = info["version"]
 
         except Exception:
             self.close()
@@ -458,7 +481,10 @@ class LegacyRemoteRepository:
                         unpacked = self.responses.pop(waiting_for[0])
                         waiting_for.pop(0)
                         handle_error(unpacked)
-                        yield unpacked[RESULT]
+                        result = unpacked[RESULT]
+                        if self.borg1_server and cmd in BYTES_RESULT_METHODS:
+                            result = bytes_from_borg1(result)
+                        yield result
                         if not waiting_for and not calls:
                             return
                     except KeyError:
@@ -614,7 +640,9 @@ class LegacyRemoteRepository:
 
     def close(self):
         if self.p:
-            self.call("close", {}, wait=True)
+            if not self.borg1_server:
+                # borg 1.x servers close the repository when stdin gets closed, they have no close method.
+                self.call("close", {}, wait=True)
             self.p.stdin.close()
             self.p.stdout.close()
             self.p.wait()
@@ -624,9 +652,14 @@ class LegacyRemoteRepository:
         for resp in self.call_many("async_responses", calls=[], wait=True, async_wait=wait):
             return resp
 
-    @api(since=parse_version("2.0.0b8"))
     def get_manifest(self):
-        """actual remoting is done via self.call in the @api decorator"""
+        if not self.borg1_server:
+            return self.call("get_manifest", {})
+        # borg 1.x servers have no get_manifest method, so get the manifest object like LegacyRepository does.
+        try:
+            return self.get(Manifest.MANIFEST_ID)
+        except LegacyRepository.ObjectNotFound:
+            raise NoManifestError
 
     @api(since=parse_version("2.0.0b8"))
     def put_manifest(self, data):